test: swap bria into tx-display tests
